feat(groth16): rename feature to r1cs
