+++ src/lib.rs
@@ -19,8 +19,8 @@ pub use encoding::Encoding;
 pub use error::EncodingError;
 pub use field_ext::FieldExt;
 
-#[cfg(feature = "groth16")]
-pub mod groth16;
+#[cfg(feature = "r1cs")]
+pub mod r1cs;
 
 pub use ark_bls12_377::Bls12_377;
 pub use ark_ed_on_bls12_377::{Fq, Fr};
